Move CreditCardFields to own file
